+++ running-races-ui/src/app/features/ub/components/map/map.ts
@@ -7,6 +7,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterLink } from '@angular/router';
 import { Router } from '@angular/router';
+import { MapService } from '../../services/map';
 
 @Component({
   selector: 'app-map',
@@ -20,8 +21,6 @@ export class MapComponent implements OnInit, OnDestroy {
   private sectionService = inject(SectionService);
   private map!: L.Map;
 
-  private readonly shadowUrl = 'https://unpkg.com/leaflet@1.9.4/dist/images/marker-shadow.png';
-  private readonly iconUrl = 'https://unpkg.com/leaflet@1.9.4/dist/images/marker-icon.png';
   sections: Section[] = [];
 
   ngOnInit(): void {
@@ -35,20 +34,10 @@ export class MapComponent implements OnInit, OnDestroy {
     }
   }
 
+  private mapService = inject(MapService);
 
   private initMap(): void {
-   
-    L.Marker.prototype.options.icon = L.icon({
-      iconUrl: this.iconUrl,
-      shadowUrl: this.shadowUrl,
-      iconSize: [10, 16],
-      iconAnchor: [5, 16]
-    });
-
-    this.map = L.map('map').setView([46.87, 17.73], 10);
-    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      attribution: '© OpenStreetMap contributors'
-    }).addTo(this.map);
+    this.map = this.mapService.initMap('map', [46.87, 17.73], 10);
   }
 
   private loadSections(): void {
@@ -61,12 +50,7 @@ export class MapComponent implements OnInit, OnDestroy {
   private drawSections(): void {
     this.sections.forEach(section => {
       const isStart = section.order === 1;
-      const startIcon = L.icon({
-        iconUrl: 'https://raw.githubusercontent.com/pointhi/leaflet-color-markers/master/img/marker-icon-2x-green.png',
-        shadowUrl: this.shadowUrl,
-        iconSize: [20, 33],
-        iconAnchor: [10, 33]
-      });
+      const icon = isStart ? this.mapService.getStartIcon() : L.Marker.prototype.options.icon;
       const start = section.startWayPoint;
       const end = section.endWayPoint;
       if (start?.lat && start?.lng && end?.lat && end?.lng) {
@@ -75,11 +59,11 @@ export class MapComponent implements OnInit, OnDestroy {
           { color: '#078080', weight: 3 }
         ).addTo(this.map);
 
-        L.marker([start.lat, start.lng], {
-          title: start.name,
-          icon: isStart ? startIcon : L.Marker.prototype.options.icon
-        })
-          .bindPopup(`<b>${start.name}</b><br>${section.name} (${section.distance} km)`)
+        this.mapService.createMarker(
+          start.lat, start.lng, start.name,
+          `<b>${start.name}</b><br>${section.name} (${section.distance} km)`,
+          section.order === 1
+        )
           .on('dblclick', () => this.router.navigate(['/ub/waypoints'], { queryParams: { search: start.name } }))
           .addTo(this.map);
       }